still on deleting threads...but if you didnt create the thread, you cannot delete it

+++ app/Http/Controllers/ThreadsController.php
@@ -120,6 +120,13 @@ class ThreadsController extends Controller
      */
     public function destroy($channels, Thread $thread)
     {
+        if($thread->user_id != auth()->id()){
+            if(request()->wantsJson()){
+                return response(['status' => 'Permission Denied'], 403); 
+            }
+
+            return redirect('/login');
+        }
         
         $thread->delete();
